+++ core/appmodule/genesis.go
@@ -2,6 +2,7 @@ package appmodule
 
 import (
 	"context"
+	"encoding/json"
 	"io"
 
 	"cosmossdk.io/core/appmodule/v2"
@@ -27,6 +28,14 @@ type HasGenesisAuto interface {
 
 	// ExportGenesis exports module state to the genesis target.
 	ExportGenesis(context.Context, GenesisTarget) error
+}
+
+// HasGenesisBasics is the legacy interface for stateless genesis methods.
+type HasGenesisBasics interface {
+	HasName
+
+	DefaultGenesis() json.RawMessage
+	ValidateGenesis(json.RawMessage) error
 }
 
 // GenesisSource is a source for genesis data in JSON format. It may abstract over a
