+++ runtime/v2/builder.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io"
 
+	"cosmossdk.io/core/appmodule"
 	appmodulev2 "cosmossdk.io/core/appmodule/v2"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/core/transaction"
@@ -37,7 +38,7 @@ func (a *AppBuilder) DefaultGenesis() map[string]json.RawMessage {
 // This is the primary hook for integrating with modules which are not registered using the app config.
 func (a *AppBuilder) RegisterModules(modules ...appmodulev2.AppModule) error {
 	for _, appModule := range modules {
-		if mod, ok := appModule.(appmodulev2.HasName); ok {
+		if mod, ok := appModule.(appmodule.HasName); ok {
 			name := mod.Name()
 			if _, ok := a.app.moduleManager.modules[name]; ok {
 				return fmt.Errorf("module named %q already exists", name)
@@ -48,7 +49,7 @@ func (a *AppBuilder) RegisterModules(modules ...appmodulev2.AppModule) error {
 				mod.RegisterInterfaces(a.app.interfaceRegistrar)
 			}
 
-			if mod, ok := appModule.(appmodulev2.HasAminoCodec); ok {
+			if mod, ok := appModule.(appmodule.HasAminoCodec); ok {
 				mod.RegisterLegacyAminoCodec(a.app.amino)
 			}
 		}
